requirements.txt
Quart==0.17.0
gunicorn==20.1.0
uvicorn==0.22.0
aiohttp==3.8.5
asyncpg==0.24.0
celery==5.2.7
redis==4.3.4
google-api-python-client==2.39.0
python-dotenv==1.0.0
pybreaker==0.6.0
